+++ src/lib/index.js
@@ -48,9 +48,9 @@ let morning = {
         black : 'bk',
         white : 'wh'
     },
-    _devVersion : '0.12.57-2',
+    _devVersion : '0.12.58',
     isMorning : true,
-    version : '0.12.57',
+    version : '0.12.58',
     map : {}
 };